requirements.txt
# Core dependencies for Streamlit Cloud deployment
streamlit>=1.28.0
plotly>=5.17.0

# LangChain ecosystem
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.2.0

# Data processing
pandas>=2.0.0
numpy>=1.21.0,<1.28.0
scipy>=1.11.0

# Excel support
openpyxl>=3.1.0

# HTTP requests
requests>=2.28.0

# Environment variables
python-dotenv>=1.0.0

# Data types
pydantic>=2.7.0
typing-extensions>=4.9.0

# Network analysis
networkx>=3.2.0

# Finance data
yfinance>=0.2.18

# Additional dependencies that might be needed
httpx>=0.28.0
anyio>=4.0.0

# Testing and coverage
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
coverage>=7.3.0